@@ -4,7 +4,7 @@ import android.app.Activity
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
-import android.support.v4.app.ActivityCompat
+import androidx.core.app.ActivityCompat
 import com.markodevcic.peko.rationale.PermissionRationale
 import java.util.concurrent.atomic.AtomicReference
 
